Implement timeouts on expensive git calls
So that this can remain in our universal $PS1 even when we're
navigating very slow remote filesystems (eg, sshfs, where even lookups
to see whether ".git" exists might take seconds), we need to implement
a couple of drastic timeouts.  For those calls where we just want a
success/failure ("are we in a git directory"), don't try to wait for
and attempt to parse stdout from a command since that has to wait for
all children even if the foreground timeout process has sent the kill
signal to the child, whereas "timeout -s 9 0.1 $cmd" leaves the child
in the background, and we can know within 0.1 second that it failed
(we know we're on a very slow filesystem if it can't succesfully find
".git" somewhere in the heirarchy within 0.1 seconds).

Once we know we're somewhere in a git repo, other git operations might
still take a very long time to respond if the tree is big enough.
Give those calls a second to respond before we too time them out.
Perhaps they'll be quicker when the PS1 returns next time, because it
will have cached more of the filesystem.  But don't want to slow
things down excessively forever...