speed up very slow filesystems: remove seemingly redundant "-r ${variable}x" eval
-n ${variable}x always expands to x at minimum, so this check seems
flawed, and then "-r x" check ends up taking a very long time on very
slow remote filesystems like sshfs.  It looks like it's trying to
protect against unset __PROMPT_COLORS_FILE, but the checks work just
fine even if that was unset (-r "" is false).  Upon removal of that
extraneous x, I couldn't find anything that failed, and now we're not
waiting many seconds for the prompt to return